chore: Code cleanups (#375)
- Replace org.assertj.core.util.Lists.emptyList() by
Collections.emptyList()
- Remove unnecessary casts
- Remove unused imports
- Use switch expressions
- pom.xml
  - remove redundant version for clirr-plugin

+++ engine/src/main/java/org/operaton/bpm/engine/impl/variable/serializer/IntegerValueSerializer.java
@@ -40,7 +40,7 @@ public class IntegerValueSerializer extends PrimitiveValueSerializer<IntegerValu
     Integer value = variableValue.getValue();
 
     if (value!=null) {
-      valueFields.setLongValue(((Integer) value).longValue());
+      valueFields.setLongValue(value.longValue());
       valueFields.setTextValue(value.toString());
     } else {
       valueFields.setLongValue(null);
