+++ engine/src/test/java/org/operaton/bpm/engine/test/bpmn/event/compensate/helper/UndoService.java
@@ -33,7 +33,7 @@ public class UndoService implements JavaDelegate {
     String variableName = (String) counterName.getValue(execution);
     Object variable = execution.getVariable(variableName);
     if(variable == null) {
-      execution.setVariable(variableName, (Integer) 1);
+      execution.setVariable(variableName, 1);
     }
     else  {
       execution.setVariable(variableName, ((Integer)variable)+1);
